Make the game be finished
Add an story for the user that play this game.
Make the game be finished by adding an end.
Add a counter to give the game a sense of urgency because the story asks
for it.
Add sit action. Player can sit only in the toilet.
Add rooms and items need to complet the game.
- Refactor Run method from World.
- Refactor other methods.